+++ packages/viz/src/components/settings/Diabeloop.js
@@ -50,12 +50,15 @@ function renderDiabeloopParameters(parametersByLevel) {
     {
       key: 'name',
       label: t('Parameter'),
+      className: 'secondaryLabelWithMainBold',
     }, {
       key: 'value',
       label: t('Value'),
+      className: 'secondaryLabelWithMainBold',
     }, {
       key: 'unit',
       label: t('Unit'),
+      className: 'secondaryLabelWithMainBold',
     },
   ];
 
@@ -63,7 +66,7 @@ function renderDiabeloopParameters(parametersByLevel) {
   parametersByLevel.forEach((parameters, level) => {
     const title = {
       label: {
-        main: `${t('level')} ${level}`,
+        main: level === 1 ? '' : `${t('Advanced')}`,
       },
       className: styles.bdlgSettingsHeader,
     };